+++ components/domains/index.ts
@@ -5,4 +5,5 @@ export { default as Thumbnail } from "./Thumbnail";
 export { default as LongPress } from "./LongPress";
 export { default as PlaylistCard } from "./PlaylistCard";
 export { default as AddSongScreen } from "./AddSongScreen";
+export { default as MoodScreen } from "./MoodScreen";
 export { default as ChangeMoodScreen } from "./ChangeMoodScreen";
